Move utils to different modules
